mcp/requirements.txt
fastapi
fastmcp
uvicorn
rav
python-dotenv
requests
pypdf